changed variable names for simplicity in access_census.R
created ArcGIS-friendly wide-format shapfiles

read shapefiles in to ArcGIS, created a file geodatabase, projected shapefiles to match landcover data, summarized landcover data, and joined summary tables back to feature class. Each geoprocessing step was written in to a non-reproducible *.py file. It is not stand alone becuase it depends on speciality arcMap toolboxes (which are provided on git).